renamed swarm to batch
